SERVICE/backend/requirements.txt
fastapi~=0.115.6
pydanti~=2.10.4
pydantic-settings~=2.7.0
uvicorn~=0.34.0
scikit-learn~=1.6.0
pandas~=2.2.3
python-multipart~=0.0.20